Move generator blocks into new folder
