+++ packages/demo/src/configuration/blocks/generators/auroraBlock.fragment.glsl
