+++ packages/demo/src/configuration/blocks/generators/fireworksBlock.fragment.glsl
