+++ packages/demo/src/configuration/blocks/generators/starryPlanesBlock.fragment.glsl
