+++ packages/demo/src/configuration/editor/blockEditorRegistrations.ts
@@ -12,10 +12,10 @@ import { GreenScreenBlock } from "../blocks/effects/greenScreenBlock";
 import { KaleidoscopeBlock } from "../blocks/effects/kaleidoscopeBlock";
 import { PixelateBlock } from "../blocks/effects/pixelateBlock";
 import { PosterizeBlock } from "../blocks/effects/posterizeBlock";
-import { StarryPlanesBlock } from "../blocks/effects/starryPlanesBlock";
-import { TunnelBlock } from "../blocks/effects/tunnelBlock";
-import { FireworksBlock } from "../blocks/effects/fireworksBlock";
-import { AuroraBlock } from "../blocks/effects/auroraBlock";
+import { StarryPlanesBlock } from "../blocks/generators/starryPlanesBlock";
+import { TunnelBlock } from "../blocks/generators/tunnelBlock";
+import { FireworksBlock } from "../blocks/generators/fireworksBlock";
+import { AuroraBlock } from "../blocks/generators/auroraBlock";
 import { VhsGlitchBlock } from "../blocks/effects/vhsGlitchBlock";
 import { SketchBlock } from "../blocks/effects/sketchBlock";
 import { SoftThresholdBlock } from "../blocks/effects/softThresholdBlock";
